py/requirements.txt
python-dateutil
numpy
pyvisa-py
h5py
pyserial
plotly
scipy
fastapi
uvicorn[standard]
pymodbus